Move the review mode tabs into the sidebar
